Change FormExtractor so it can be extended
I'm using MopaBootstrapBundle which has some extra options in forms for help text which need to be translated and extracted. It would be easiest for me to create a new extractor which extends the current form extractor but because the methods are private, I can't reuse a lot of the code.