Allow skipping dirs validation during configuration warmup
There is no reason to validate the dirs until we actually run the
extraction. This PR adds the validate_dir option which defaults to
`true` so users can opt for not validating the dirs paths during
configuration warmup phase.

It's possible that we still want to do validation at extract phase
and we can add the code to do that if it is necessary.